[DPP-1219] Organize packages in participant-local-store Bazel package
changelog_begin
changelog_end
