+++ ledger/ledger-api-auth/src/main/scala/com/digitalasset/ledger/api/auth/UserRightsChangeAsyncChecker.scala
@@ -9,7 +9,7 @@ import java.util.concurrent.atomic.AtomicReference
 import akka.actor.Scheduler
 import com.daml.ledger.api.auth.interceptor.AuthorizationInterceptor
 import com.daml.ledger.api.domain
-import com.daml.ledger.participant.state.index.v2.UserManagementStore
+import com.daml.ledger.participant.localstore.api.UserManagementStore
 import com.daml.lf.data.Ref
 import com.daml.logging.LoggingContext
