+++ ledger/ledger-api-auth/src/test/suite/scala/com/digitalasset/ledger/api/auth/AuthorizerSpec.scala
@@ -11,7 +11,7 @@ import org.scalatest.matchers.should.Matchers
 import java.time.Instant
 
 import com.daml.ledger.api.testing.utils.AkkaBeforeAndAfterAll
-import com.daml.ledger.participant.state.index.v2.UserManagementStore
+import com.daml.ledger.participant.localstore.api.UserManagementStore
 import com.daml.logging.LoggingContext
 import org.mockito.MockitoSugar
