+++ ledger/participant-integration-api/src/main/scala/platform/apiserver/update/UserUpdateMapper.scala
@@ -5,7 +5,7 @@ package com.daml.platform.apiserver.update
 
 import com.daml.ledger.api.domain
 import com.daml.ledger.api.domain.User
-import com.daml.ledger.participant.state.index.v2._
+import com.daml.ledger.participant.localstore.api.{ObjectMetaUpdate, UserUpdate}
 import com.daml.lf.data.Ref
 
 object UserUpdateMapper extends UpdateMapperBase {
