+++ ledger/participant-integration-api/src/test/suite/scala/db/migration/translation/apiserver/services/admin/ApiPartyManagementServiceSpec.scala
@@ -11,10 +11,10 @@ import com.daml.ledger.api.domain.ParticipantParty.PartyRecord
 import com.daml.ledger.api.domain.{ObjectMeta, PartyDetails, PartyEntry}
 import com.daml.ledger.api.testing.utils.AkkaBeforeAndAfterAll
 import com.daml.ledger.api.v1.admin.party_management_service.AllocatePartyRequest
+import com.daml.ledger.participant.localstore.api.PartyRecordStore
 import com.daml.ledger.participant.state.index.v2.{
   IndexPartyManagementService,
   IndexTransactionsService,
-  PartyRecordStore,
 }
 import com.daml.ledger.participant.state.{v2 => state}
 import com.daml.lf.data.Ref
