+++ ledger/participant-integration-api/src/test/suite/scala/platform/apiserver/update/PartyRecordUpdateMapperSpec.scala
@@ -5,7 +5,7 @@ package com.daml.platform.apiserver.update
 
 import com.daml.ledger.api.domain.ParticipantParty.PartyDetails
 import com.daml.ledger.api.domain.ObjectMeta
-import com.daml.ledger.participant.state.index.v2.{ObjectMetaUpdate, PartyDetailsUpdate}
+import com.daml.ledger.participant.localstore.api.{ObjectMetaUpdate, PartyDetailsUpdate}
 import com.daml.lf.data.Ref
 import com.google.protobuf.field_mask.FieldMask
 import org.scalatest.EitherValues
