+++ ledger/participant-local-store/src/main/scala/com/daml/ledger/participant/localstore/CachedUserManagementStore.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.usermanagement
+package com.daml.ledger.participant.localstore
 
 import java.time.Duration
 import java.util.concurrent.{CompletableFuture, Executor}
@@ -9,8 +9,8 @@ import java.util.concurrent.{CompletableFuture, Executor}
 import com.daml.caching.CaffeineCache
 import com.daml.ledger.api.domain
 import com.daml.ledger.api.domain.User
-import com.daml.ledger.participant.state.index.v2.UserManagementStore.{Result, UserInfo}
-import com.daml.ledger.participant.state.index.v2.{UserManagementStore, UserUpdate}
+import com.daml.ledger.participant.localstore.api.{UserManagementStore, UserUpdate}
+import com.daml.ledger.participant.localstore.api.UserManagementStore.{Result, UserInfo}
 import com.daml.lf.data.Ref
 import com.daml.lf.data.Ref.UserId
 import com.daml.logging.LoggingContext
