+++ ledger/participant-local-store/src/main/scala/com/daml/ledger/participant/localstore/InMemoryPartyRecordStore.scala
@@ -1,20 +1,23 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.index.impl.inmemory
+package com.daml.ledger.participant.localstore
 
 import com.daml.ledger.api.domain
 import com.daml.ledger.api.domain.{ObjectMeta, ParticipantParty}
-import com.daml.ledger.participant.state.index.{LocalAnnotationsUtils, ResourceAnnotationValidation}
-import com.daml.ledger.participant.state.index.v2.PartyRecordStore.{
+import com.daml.ledger.participant.localstore.api.{
+  LedgerPartyExists,
+  PartyRecordStore,
+  PartyRecordUpdate,
+}
+import com.daml.ledger.participant.localstore.api.PartyRecordStore.{
   MaxAnnotationsSizeExceeded,
   PartyRecordExistsFatal,
   Result,
 }
-import com.daml.ledger.participant.state.index.v2.{
-  LedgerPartyExists,
-  PartyRecordStore,
-  PartyRecordUpdate,
+import com.daml.ledger.participant.localstore.utils.{
+  LocalAnnotationsUtils,
+  ResourceAnnotationValidation,
 }
 import com.daml.lf.data.Ref
 import com.daml.lf.data.Ref.Party
