+++ ledger/participant-local-store/src/main/scala/com/daml/ledger/participant/localstore/InMemoryUserManagementStore.scala
@@ -1,12 +1,15 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.index.impl.inmemory
+package com.daml.ledger.participant.localstore
 
 import com.daml.ledger.api.domain.{ObjectMeta, User, UserRight}
-import com.daml.ledger.participant.state.index.{LocalAnnotationsUtils, ResourceAnnotationValidation}
-import com.daml.ledger.participant.state.index.v2.{UserManagementStore, UserUpdate}
-import com.daml.ledger.participant.state.index.v2.UserManagementStore._
+import com.daml.ledger.participant.localstore.api.{UserManagementStore, UserUpdate}
+import com.daml.ledger.participant.localstore.api.UserManagementStore._
+import com.daml.ledger.participant.localstore.utils.{
+  LocalAnnotationsUtils,
+  ResourceAnnotationValidation,
+}
 import com.daml.lf.data.Ref
 import com.daml.lf.data.Ref.UserId
 import com.daml.logging.LoggingContext
