+++ ledger/participant-local-store/src/main/scala/com/daml/ledger/participant/localstore/api/PartyRecordStore.scala
@@ -1,10 +1,10 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.index.v2
+package com.daml.ledger.participant.localstore.api
 
 import com.daml.ledger.api.domain
-import com.daml.ledger.participant.state.index.ResourceAnnotationValidation
+import com.daml.ledger.participant.localstore.utils.ResourceAnnotationValidation
 import com.daml.lf.data.Ref
 import com.daml.logging.LoggingContext
