+++ ledger/participant-local-store/src/main/scala/com/daml/ledger/participant/localstore/utils/ResourceAnnotationValidation.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.participant.state.index
+package com.daml.ledger.participant.localstore.utils
 
 import java.nio.charset.StandardCharsets
