+++ ledger/participant-local-store/src/test/lib/com/daml/ledger/participant/localstore/PartyRecordStoreSpecBase.scala
@@ -1,9 +1,9 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.store.platform.partymanagement
+package com.daml.ledger.participant.localstore
 
-import com.daml.ledger.participant.state.index.v2.PartyRecordStore
+import com.daml.ledger.participant.localstore.api.PartyRecordStore
 import com.daml.ledger.resources.TestResourceContext
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.{Assertion, AsyncTestSuite, EitherValues, OptionValues}
