+++ ledger/participant-local-store/src/test/lib/com/daml/ledger/participant/localstore/UserStoreTests.scala
@@ -1,19 +1,19 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.store.platform.usermanagement
+package com.daml.ledger.participant.localstore
 
 import com.daml.ledger.api.domain.{ObjectMeta, User, UserRight}
-import com.daml.ledger.participant.state.index.v2.{
-  ObjectMetaUpdate,
-  UserManagementStore,
-  UserUpdate,
-}
-import com.daml.ledger.participant.state.index.v2.UserManagementStore.{
+import com.daml.ledger.participant.localstore.api.UserManagementStore.{
   MaxAnnotationsSizeExceeded,
   UserExists,
   UserNotFound,
   UsersPage,
+}
+import com.daml.ledger.participant.localstore.api.{
+  ObjectMetaUpdate,
+  UserManagementStore,
+  UserUpdate,
 }
 import com.daml.lf.data.Ref
 import com.daml.lf.data.Ref.{Party, UserId}
