+++ ledger/participant-local-store/src/test/suite/com/daml/ledger/participant/localstore/InMemoryPartyRecordStoreSpec.scala
@@ -1,11 +1,9 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.partymanagement
+package com.daml.ledger.participant.localstore
 
-import com.daml.ledger.participant.state.index.impl.inmemory.InMemoryPartyRecordStore
-import com.daml.ledger.participant.state.index.v2.PartyRecordStore
-import com.daml.platform.store.platform.partymanagement.PartyRecordStoreTests
+import com.daml.ledger.participant.localstore.api.PartyRecordStore
 import org.scalatest.freespec.AsyncFreeSpec
 
 class InMemoryPartyRecordStoreSpec extends AsyncFreeSpec with PartyRecordStoreTests {
